added back chain's default policy
The default policies of the INPUT/OUTPUT/FORWARD chains
got lost during refactoring. We now added them back. This doesn't
change anything for the examples, since they either already have
an explicit default rule or implement everything we are verifying
without relying on the default policy.

Currently, the default policy must be specified manually. The ML
parser will be able to fill this out automatically in the future.

This commit restores the behavior documented in the FM'15 paper,
footnote 4.