Fixed goto actions
iptables supports a goto target; the Semantics doesn't. In the `iptables
-L -n' format, goto appeared as a normal jump (Call) action, the fact
that it was a goto was disguised as a normal match condition.
Example:
logflags   tcp  --  0.0.0.0/0            0.0.0.0/0           [goto]  tcp
flags:0x3F/0x29

The fix is simple: the python and SML tokenizer fail if a goto action
occurs.

The only firewall where gotos occurred was SQRL_Shorewall. In this
ruleset, the behavior of goto was identical to the behavior of call.:

reject

  no RETURNs

  when called with goto, it is the last rule in the chain, afterwards,
the default policy applies

  behavior is identical to Call

smurflog, logflags

  ends with explicit deny-all, thus identical to being called

Therefore, the analysis was not wrong. For correctness, we remove it
anyway from the examples. It is available in the goto-branch.