debugging performance problem: extract_IPSets
it seems that the runtime of isballe build with all
IPPart Examples has increased by a factor of three.
This happend with commit 4df28432be8f53a106efcbf0d3e72c9f3bfa2fcc
This commit mainly changed the definition of extract_IPSets.
In artificial test, the new definition was faster.
This commit reverst only this function to the old version.
Let's see what happens to the runtime.
Runtime is commented in github commit comments.