deleted the ML files
We hacky-parsed an iptables -L -n dump to ML and exported the
isabelle-generated code to ML and loaded the whole thing again into
easabelle (hack, hack hack) to execute it.

Finally, we have nice parsers and a working haskell tool :-)
Time to throw away the legacy ML artifacts now.

By the way, the ML things was still faster than the tests with "by eval"
we do now. But the "by eval" tests are more reliable and the haskell
tool has a comparable performance w.r.t. the ML approach.